SimilarityCheck/requirements.txt
requests>=2.25.1
azure-storage-blob>=12.19.0
python-dotenv